requirements.txt
dj-database-url==0.4.2
Django==1.10.4
django-chartit==0.2.8
gunicorn==19.6.0
whitenoise==3.2.3
